+++ WireNetwork/Sources/WireNetwork/Network/NetworkStack/NetworkStack.swift
@@ -133,87 +133,3 @@ private enum NetworkState {
     case ready(rest: NetworkService, webSocket: NetworkService)
 
 }
-
-// TODO: move to network service
-private extension NetworkService {
-
-    static func makeServices(
-        backendConfig: BackendEnvironment2.Config,
-        minTLSVersion: TLSVersion,
-        proxyCredentials: ProxyCredentials? = nil
-    ) throws(InitializationError) -> (
-        rest: NetworkService,
-        webSocket: NetworkService
-    ) {
-        let proxySettings: ProxySettings?
-        if let proxyConfig = backendConfig.proxyConfig {
-            if proxyConfig.needsAuthentication {
-                guard let proxyCredentials else {
-                    throw .proxyCredentialsRequired
-                }
-                proxySettings = .authenticated(
-                    host: proxyConfig.host,
-                    port: proxyConfig.port,
-                    username: proxyCredentials.username,
-                    password: proxyCredentials.password
-                )
-            } else {
-                proxySettings = .unauthenticated(
-                    host: proxyConfig.host,
-                    port: proxyConfig.port
-                )
-            }
-        } else {
-            proxySettings = nil
-        }
-
-        let configFactory = URLSessionConfigurationFactory(
-            minTLSVersion: minTLSVersion,
-            proxySettings: proxySettings
-        )
-
-        let restService = NetworkService(
-            baseURL: backendConfig.endpoints.restAPIURL,
-            serverTrustValidator: ServerTrustValidator(
-                pinnedKeys: backendConfig.pinnedKeys,
-                currentDateProvider: .system
-            )
-        )
-
-        let restSession = URLSession(
-            configuration: configFactory.makeRESTAPISessionConfiguration(),
-            delegate: restService,
-            delegateQueue: nil
-        )
-
-        restService.configure(with: restSession)
-
-        let webSocketService = NetworkService(
-            baseURL: backendConfig.endpoints.websocketURL,
-            serverTrustValidator: ServerTrustValidator(
-                pinnedKeys: backendConfig.pinnedKeys,
-                currentDateProvider: .system
-            )
-        )
-
-        let webSocketSession = URLSession(
-            configuration: configFactory.makeWebSocketSessionConfiguration(),
-            delegate: webSocketService,
-            delegateQueue: nil
-        )
-
-        webSocketService.configure(with: webSocketSession)
-
-        return (
-            rest: restService,
-            webSocket: webSocketService
-        )
-    }
-
-    enum InitializationError: Error {
-
-        case proxyCredentialsRequired
-
-    }
-
-}
